Work on geometry selection editor.

+++ libtitania-x3d/Titania/X3D/Components/Grouping/X3DTransformNode.h
@@ -131,7 +131,7 @@ public:
 
 	virtual
 	void
-	setMatrixWithCenter (const Matrix4d &, const Vector3f &);
+	setMatrixWithCenter (const Matrix4d &, const Vector3d &);
 
 	virtual
 	void
